[fix] Character_StatsData 테이블 타입값 변경
BigInteger 타입변경으로 인한 참조 클래스 타입 변경

+++ Assets/2. Scripts/UI/CharactorPopup.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Numerics;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
@@ -60,7 +61,7 @@ public class CharactorPopup : BaseUI
         UIManager.Instance.OpenPanel(statUpgradePopup);
     }
 
-    void UpdateExp(float currentExp, double maxExp)
+    void UpdateExp(BigInteger currentExp, BigInteger maxExp)
     {
         expSlider.value = (float)(currentExp / maxExp);
